requirements.txt
Django>=4.2.0
psycopg2-binary>=2.9.9
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
python-dotenv>=1.0.0
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0 
